Removed ability to select gromacs build; re-added get_name to kernels
So apparently we can't do what we wanted here because the Kernel class
in radical.ensemblemd.kernel needs to call `get_name` for the classes in
the module registry to populate its own list of them. Note that tis
Kernel class should not be confused with the Kernel class in
radical.ensemblemd.kernel_plugins.kernel_base, which is the actual base
class for these Kernels.

This is all a bit confusing; what's the reason for this?